Develop

+++ index.php
@@ -3,7 +3,7 @@ $PageSecurity = 0;
 
 include('includes/session.inc');
 
-if ((filesize('install/InitialScripts.txt') > 0) and isset($_SESSION['DatabaseName'])) {
+if (file_exists('install/InitialScripts.txt') and (filesize('install/InitialScripts.txt') > 0) and isset($_SESSION['DatabaseName'])) {
 	echo '<meta http-equiv="refresh" content="0; url=' . $RootPath . '/InitialScripts.php">';
 	exit;
 }
